Moved Scanner, Parser, and Interpreter test to own folder
